requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
astroid==3.3.8
    # via pylint
autopep8==2.3.2
    # via -r requirements.in
cffi==1.17.1
    # via cryptography
cryptography==44.0.0
    # via -r requirements.in
dill==0.3.9
    # via pylint
isort==5.13.2
    # via pylint
mccabe==0.7.0
    # via pylint
mypy==1.14.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
platformdirs==4.3.6
    # via pylint
pycodestyle==2.12.1
    # via autopep8
pycparser==2.22
    # via cffi
pylint==3.3.3
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
tomlkit==0.13.2
    # via pylint
types-pyyaml==6.0.12.20241230
    # via -r requirements.in
typing-extensions==4.12.2
    # via mypy
